增加 Update 和 Delete 操作的检查，当没有指定更新字段和 Where 条件时，将抛出异常，防止对整表的误操作

+++ MyStaging/Helpers/DeleteBuilder.cs
@@ -30,17 +30,17 @@ namespace MyStaging.Helpers
             {
                 foreach (var item in WhereExpressionList)
                 {
-                    //PgSqlExpression expression = new PgSqlExpression();
-                    //expression.ExpressionCapture(item.Body);
-                    //WhereList.Add(expression.CommandText.ToString().ToLower());
-                    //ParamList.AddRange(expression.Parameters);
-
                     DbExpressionVisitor expression = new DbExpressionVisitor();
                     expression.Visit(item.Body);
                     WhereList.Add(expression.SqlText.Builder.ToString().ToLower());
                     ParamList.AddRange(expression.SqlText.Parameters);
                 }
             }
+
+            if (this.WhereList.Count == 0)
+                throw new ArgumentException("The delete operation must specify where conditions!");
+
+
             string cmdText = $"DELETE FROM {tableName} {"WHERE " + string.Join("\nAND ", WhereList)}";
             return base.ExecuteNonQuery(cmdText);
         }
